+++ src/blueticks/resources/__init__.py
@@ -5,7 +5,6 @@ from blueticks.resources.chats import ChatsResource
 from blueticks.resources.contacts import ContactsResource
 from blueticks.resources.engines import EnginesResource
 from blueticks.resources.groups import GroupsResource
-from blueticks.resources.messages import MessagesResource
 from blueticks.resources.newsletters import NewslettersResource
 from blueticks.resources.ping import PingResource
 from blueticks.resources.scheduled_messages import ScheduledMessagesResource
@@ -20,7 +19,6 @@ __all__ = [
     "ContactsResource",
     "EnginesResource",
     "GroupsResource",
-    "MessagesResource",
     "NewslettersResource",
     "PingResource",
     "ScheduledMessagesResource",